pyproject.toml
[project]
name = "ecoroute-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "litellm>=1.80.10",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "polyline>=2.0.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "streamlit-folium>=0.25.3",
]
